+++ include/Stack.h
@@ -7,8 +7,8 @@
 typedef genVec Stack;
 
 
-Stack* stack_create(u32 n, u16 data_size, genVec_copy_fn copy_fn, genVec_move_fn move_fn, genVec_delete_fn del_fn);
-Stack* stack_create_val(u32 n, const u8* val, u16 data_size, genVec_copy_fn copy_fn, genVec_move_fn move_fn,
+Stack* stack_create(u64 n, u32 data_size, genVec_copy_fn copy_fn, genVec_move_fn move_fn, genVec_delete_fn del_fn);
+Stack* stack_create_val(u64 n, const u8* val, u32 data_size, genVec_copy_fn copy_fn, genVec_move_fn move_fn,
                         genVec_delete_fn del_fn);
 
 void stack_destroy(Stack* stk);
@@ -21,7 +21,7 @@ void      stack_pop(Stack* stk, u8* popped);
 void      stack_peek(Stack* stk, u8* peek);
 const u8* stack_peek_ptr(Stack* stk);
 
-static inline u32 stack_size(Stack* stk)
+static inline u64 stack_size(Stack* stk)
 {
     return genVec_size(stk);
 }
@@ -31,7 +31,7 @@ static inline u8 stack_empty(Stack* stk)
     return genVec_empty(stk);
 }
 
-static inline u32 stack_capacity(Stack* stk)
+static inline u64 stack_capacity(Stack* stk)
 {
     return genVec_capacity(stk);
 }
